Add service tests
